refactor(properties): move selector logic into managers and instance methods
